move method invoker to service module
