+++ modules/service/src/main/java/org/mule/runtime/module/service/internal/manager/MethodInvoker.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2023 Salesforce, Inc. All rights reserved.
  */
-package org.mule.runtime.core.internal.util;
+package org.mule.runtime.module.service.internal.manager;
 
 import java.lang.reflect.Method;
